Update trivial-download.lisp
fix the `(unless quiet` in `with-download` macro. The original version happens to be right because the same name `quiet` in the definition of `download` function, but will come to error if directly use `with-download` with `:quiet`. Add another option `:drakma-options` that may be helpful to send additional information when download, e.g. a fake header, cookies or user agent that may be helpful to download files in some websites.